\name{nearPPSD-package}
\alias{nearPPSD-package}
\alias{nearPPSD}
\docType{package}
\title{
\packageTitle{nearPPSD}
}
\description{
\packageDescription{nearPPSD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nearPPSD}
\packageIndices{nearPPSD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nearPPSD}

Maintainer: \packageMaintainer{nearPPSD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
