% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearPPSD}
\alias{nearPPSD}
\title{Finding the nearest patterned positive-semidefinite matrix}
\usage{
nearPPSD(X, eigenTol = 1e-06, convTol = 1e-07, psdTol = 1e-08,
  maxit = 100L)
}
\arguments{
\item{X}{a numeric matrix.}

\item{eigenTol}{defines relative positiveness of eigenvalues compared to largest one.}

\item{convTol}{convergence tolerance.}

\item{psdTol}{tolerance for enforcing positive definiteness.}

\item{maxit}{maximum number of iterations allowed.}
}
\value{
A list of five elements.

\item{X}{The resulting patterned positive-semidefinite matrix.}
\item{iteration}{Number of iterations used.}
\item{convergence}{Whether the algorithmed converged.}
\item{normF}{Frobenius norm between the resulting optimal and the initial matrices.}
\item{normC}{Inf norm between the resulting optimal and the initial matrices.}
}
\description{
Finding the nearest patterned positive-semidefinite matrix
}
\references{
R. Escalante and M. Raydan \emph{Dykstra's Algorithm for a Constrained Least-squares Matrix Problem} Numerical Linear Algebra with Applications (1996) 459--471

H. Hu and I. Olkin \emph{A numerical procedure for finding the positive definite matrix closest to a patterned matrix} Statistics & Probability Letters (1991) 511--515

N. Higham \emph{Computing the nearest correlation matrix - a problem from finance} IMA Journal of Numerical Analysis (2002) 329--343
}
\author{
Julius Vainora
}
